refactor(proxy): extract table scan handlers
Move rnd scan callbacks, scan batch loading, and row-position re-read handling into table_scan.cc.
Register table_scan.cc in the proxy build.
