+++ proxy/ha_lineairdb.cc
@@ -1130,184 +1130,6 @@ const Item *ha_lineairdb::cond_push(const Item *cond) {
 }
 
 /**
-  @brief
-  rnd_init() is called when the system wants the storage engine to do a table
-  scan. See the lineairdb in the introduction at the top of this file to see
-  when rnd_init() is called.
-
-  @details
-  Called from filesort.cc, records.cc, sql_handler.cc, sql_select.cc,
-  sql_table.cc, and sql_update.cc.
-
-  @see
-  filesort.cc, records.cc, sql_handler.cc, sql_select.cc, sql_table.cc and
-  sql_update.cc
-*/
-int ha_lineairdb::rnd_init(bool) {
-  DBUG_ENTER("ha_lineairdb::rnd_init");
-  scanned_keys_.clear();
-  scanned_values_.clear();
-  scan_cache_.clear();
-  buffer_position_ = 0;
-  last_batch_key_.clear();
-  scan_exhausted_ = false;
-  last_fetched_primary_key_.clear();
-  current_position_ = 0;
-  stats.records = 0;
-
-  change_active_index(table->s->primary_key);
-
-  auto tx = get_transaction(ha_thd());
-
-  if (tx->is_aborted()) {
-    DBUG_RETURN(abort_errno(tx));
-  }
-
-  tx->choose_table(db_table_name);
-
-  // Predicate pushdown: propagate filter serialized by cond_push() to the
-  // transaction. When an aggregation pushdown is pending, the tx filter was
-  // prepared by tx_set_pushed_aggregate from the statement WHERE — clearing
-  // it here (cond_push is empty in practice) would make the server aggregate
-  // UNFILTERED rows.
-  if (!pushed_filter_serialized_.empty()) {
-    tx->set_pushed_filter(pushed_filter_serialized_);
-  } else if (!tx->has_pushed_aggregate()) {
-    tx->clear_pushed_filter();
-  }
-
-  if (prefetch_needs_legacy_dml_handler(ha_thd(), tx)) {
-    DBUG_RETURN(prefetch_reject_unsupported(
-        ha_thd(), tx, "legacy DML full/reverse table scan"));
-  }
-
-  // The optimizer has run, so the QEP is available.
-  // Statement-scoped autogen stages and executes the prefetch plan once per
-  // statement; an unsupported QEP fails here.
-  if (int err = maybe_prefetch_for_statement(ha_thd(), tx, table)) {
-    DBUG_RETURN(err);
-  }
-
-  DBUG_RETURN(0);
-}
-
-int ha_lineairdb::rnd_end() {
-  DBUG_TRACE;
-  // NOTE: Do NOT clear scan_cache_ / scanned_values_ here.
-  // MySQL calls rnd_end() after scanning, then rnd_pos() to re-read rows in
-  // sorted order. In InnoDB the re-read hits the Buffer Pool, but we need our
-  // own cache (scan_cache_) since re-reads would otherwise require RPCs.
-  // Cleared at the start of the next rnd_init() instead.
-  buffer_position_ = 0;
-  last_batch_key_.clear();
-  scan_exhausted_ = false;
-  blobroot.Clear();
-  return 0;
-}
-
-/**
- * @brief Fetch next batch of rows for full table scan.
- *
- * Proxy adaptation: uses get_matching_keys_and_values_from_prefix("")
- * instead of SE's Scan() callback.
- */
-bool ha_lineairdb::fetch_next_batch() {
-  DBUG_ENTER("ha_lineairdb::fetch_next_batch");
-
-  auto tx = get_transaction(ha_thd());
-  if (tx->is_aborted()) {
-    DBUG_RETURN(false);
-  }
-
-  tx->choose_table(db_table_name);
-
-  scanned_keys_.clear();
-  scanned_values_.clear();
-  scan_cache_.clear();
-  buffer_position_ = 0;
-
-  // Proxy: fetch all rows via RPC in one call (no batched Scan callback)
-  auto key_value_pairs = tx->get_matching_keys_and_values_from_prefix("");
-
-  for (auto &kv : key_value_pairs) {
-    // skip tombstones
-    if (kv.second.empty()) continue;
-
-    // Store row data and build scan_cache_ so rnd_pos() can reuse it later.
-    size_t idx = scanned_keys_.size();
-    scanned_keys_.push_back(kv.first);
-    const auto &val = kv.second;
-    scanned_values_.emplace_back(
-        reinterpret_cast<const std::byte *>(val.data()),
-        reinterpret_cast<const std::byte *>(val.data()) + val.size());
-    scan_cache_[kv.first] = idx;
-  }
-
-  // Check if tx was aborted during RPC
-  if (tx->is_aborted()) {
-    thd_mark_transaction_to_rollback(ha_thd(), 1);
-    DBUG_RETURN(false);
-  }
-
-  if (scanned_keys_.empty()) {
-    DBUG_RETURN(false);
-  }
-
-  // Mark as exhausted since we fetched everything in one call
-  scan_exhausted_ = true;
-
-  DBUG_RETURN(true);
-}
-
-/**
-  @brief
-  This is called for each row of the table scan. When you run out of records
-  you should return HA_ERR_END_OF_FILE. Fill buff up with the row information.
-  The Field structure for the table is the key to getting data into buf
-  in a manner that will allow the server to understand it.
-
-  @details
-  Called from filesort.cc, records.cc, sql_handler.cc, sql_select.cc,
-  sql_table.cc, and sql_update.cc.
-
-  @see
-  filesort.cc, records.cc, sql_handler.cc, sql_select.cc, sql_table.cc and
-  sql_update.cc
-*/
-
-// assumption: takes 1 row
-int ha_lineairdb::rnd_next(uchar *buf) {
-  DBUG_ENTER("ha_lineairdb::rnd_next");
-  ha_statistic_increment(&System_status_var::ha_read_rnd_next_count);
-
-  if (buffer_position_ >= scanned_keys_.size()) {
-    if (scan_exhausted_) {
-      DBUG_RETURN(HA_ERR_END_OF_FILE);
-    }
-
-    if (!fetch_next_batch()) {
-      auto tx = get_transaction(ha_thd());
-      if (tx->is_aborted()) {
-        DBUG_RETURN(abort_errno(tx));
-      }
-      scan_exhausted_ = true;
-      DBUG_RETURN(HA_ERR_END_OF_FILE);
-    }
-  }
-
-  auto &key = scanned_keys_[buffer_position_];
-  auto &value = scanned_values_[buffer_position_];
-  buffer_position_++;
-
-  int error = set_fields_from_lineairdb(buf, value.data(), value.size());
-  if (error == 0) {
-    last_fetched_primary_key_ = key;
-  }
-  current_position_++;
-  DBUG_RETURN(error);
-}
-
-/**
  * @brief Attach an aggregate spec and its server-side WHERE filter to the tx.
  *
  * Server aggregation folds base rows into group rows, so MySQL cannot recheck
@@ -1379,100 +1201,6 @@ bool ha_lineairdb::agg_next_raw(std::string_view *out_value) {
       reinterpret_cast<const char *>(value.data()), value.size());
   current_position_++;
   return true;
-}
-
-/**
-  @brief
-  position() is called after each call to rnd_next() if the data needs
-  to be ordered. You can do something like the following to store
-  the position:
-  @code
-  my_store_ptr(ref, ref_length, current_position);
-  @endcode
-
-  @details
-  The server uses ref to store data. ref_length in the above case is
-  the size needed to store current_position. ref is just a byte array
-  that the server will maintain. If you are using offsets to mark rows, then
-  current_position should be the offset. If it is a primary key like in
-  BDB, then it needs to be a primary key.
-
-  Called from filesort.cc, sql_select.cc, sql_delete.cc, and sql_update.cc.
-
-  @see
-  filesort.cc, sql_select.cc, sql_delete.cc and sql_update.cc
-*/
-void ha_lineairdb::position(const uchar *) {
-  DBUG_TRACE;
-
-  if (last_fetched_primary_key_.empty()) {
-    return;
-  }
-
-  store_primary_key_in_ref(last_fetched_primary_key_);
-}
-
-/**
-  @brief
-  This is like rnd_next, but you are given a position to use
-  to determine the row. The position will be of the type that you stored in
-  ref. You can use ha_get_ptr(pos,ref_length) to retrieve whatever key
-  or position you saved when position() was called.
-
-  @details
-  Called from filesort.cc, records.cc, sql_insert.cc, sql_select.cc, and
-  sql_update.cc.
-
-  @see
-  filesort.cc, records.cc, sql_insert.cc, sql_select.cc and sql_update.cc
-*/
-int ha_lineairdb::rnd_pos(uchar *buf, uchar *pos) {
-  DBUG_TRACE;
-
-  std::string primary_key = extract_primary_key_from_ref(pos);
-
-  if (primary_key.empty()) {
-    return HA_ERR_KEY_NOT_FOUND;
-  }
-
-  // Return from scan_cache_ if available (equivalent of hitting InnoDB's Buffer
-  // Pool). Without this, every re-read would be a separate RPC to LineairDB.
-  auto cache_it = scan_cache_.find(primary_key);
-  if (cache_it != scan_cache_.end()) {
-    auto &value = scanned_values_[cache_it->second];
-    if (set_fields_from_lineairdb(buf, value.data(), value.size())) {
-      return HA_ERR_OUT_OF_MEM;
-    }
-    last_fetched_primary_key_ = primary_key;
-    return 0;
-  }
-
-  auto tx = get_transaction(ha_thd());
-
-  if (tx->is_aborted()) {
-    return abort_errno(tx);
-  }
-
-  tx->choose_table(db_table_name);
-  auto result = tx->read(primary_key);
-
-  // read() aborts the tx on a prefetch cache miss; catch it before the empty
-  // result below reads as a missing key.
-  if (tx->is_aborted()) {
-    return abort_errno(tx);
-  }
-  if (result.first == nullptr || result.second == 0) {
-    return HA_ERR_KEY_NOT_FOUND;
-  }
-
-  if (set_fields_from_lineairdb(buf, result.first, result.second)) {
-    tx->set_status_to_abort();
-    return HA_ERR_OUT_OF_MEM;
-  }
-
-  last_fetched_primary_key_ = primary_key;
-
-  return 0;
 }
 
 /**
