+++ proxy/table_scan.cc
@@ -1,0 +1,207 @@
+#include "storage/lineairdb/ha_lineairdb.hh"
+
+#include <cstddef>
+#include <string>
+
+#include "lineairdb_prefetch.hh"
+#include "my_dbug.h"
+#include "sql/sql_class.h"
+#include "sql/table.h"
+
+// Handler full-table-scan entry points. These methods back MySQL's rnd_* scan
+// callbacks with LineairDB range reads and maintain the local row cache used
+// for sorted re-reads through rnd_pos().
+
+int ha_lineairdb::rnd_init(bool) {
+  DBUG_ENTER("ha_lineairdb::rnd_init");
+  scanned_keys_.clear();
+  scanned_values_.clear();
+  scan_cache_.clear();
+  buffer_position_ = 0;
+  last_batch_key_.clear();
+  scan_exhausted_ = false;
+  last_fetched_primary_key_.clear();
+  current_position_ = 0;
+  stats.records = 0;
+
+  change_active_index(table->s->primary_key);
+
+  auto tx = get_transaction(ha_thd());
+
+  if (tx->is_aborted()) {
+    DBUG_RETURN(abort_errno(tx));
+  }
+
+  tx->choose_table(db_table_name);
+
+  // Predicate pushdown: propagate filter serialized by cond_push() to the
+  // transaction. When an aggregation pushdown is pending, the tx filter was
+  // prepared by tx_set_pushed_aggregate from the statement WHERE. Clearing it
+  // here would make the server aggregate unfiltered rows.
+  if (!pushed_filter_serialized_.empty()) {
+    tx->set_pushed_filter(pushed_filter_serialized_);
+  } else if (!tx->has_pushed_aggregate()) {
+    tx->clear_pushed_filter();
+  }
+
+  if (prefetch_needs_legacy_dml_handler(ha_thd(), tx)) {
+    DBUG_RETURN(prefetch_reject_unsupported(
+        ha_thd(), tx, "legacy DML full/reverse table scan"));
+  }
+
+  // The optimizer has run, so the QEP is available.
+  // Statement-scoped autogen stages and executes the prefetch plan once per
+  // statement; an unsupported QEP fails here.
+  if (int err = maybe_prefetch_for_statement(ha_thd(), tx, table)) {
+    DBUG_RETURN(err);
+  }
+
+  DBUG_RETURN(0);
+}
+
+int ha_lineairdb::rnd_end() {
+  DBUG_TRACE;
+  // Do not clear scan_cache_ or scanned_values_ here. MySQL can call rnd_end()
+  // after scanning and then call rnd_pos() to re-read rows in sorted order.
+  // Clear them at the start of the next rnd_init() instead.
+  buffer_position_ = 0;
+  last_batch_key_.clear();
+  scan_exhausted_ = false;
+  blobroot.Clear();
+  return 0;
+}
+
+bool ha_lineairdb::fetch_next_batch() {
+  DBUG_ENTER("ha_lineairdb::fetch_next_batch");
+
+  auto tx = get_transaction(ha_thd());
+  if (tx->is_aborted()) {
+    DBUG_RETURN(false);
+  }
+
+  tx->choose_table(db_table_name);
+
+  scanned_keys_.clear();
+  scanned_values_.clear();
+  scan_cache_.clear();
+  buffer_position_ = 0;
+
+  auto key_value_pairs = tx->get_matching_keys_and_values_from_prefix("");
+
+  for (auto &kv : key_value_pairs) {
+    if (kv.second.empty()) continue;
+
+    // Store row data and build scan_cache_ so rnd_pos() can reuse it later.
+    size_t idx = scanned_keys_.size();
+    scanned_keys_.push_back(kv.first);
+    const auto &val = kv.second;
+    scanned_values_.emplace_back(
+        reinterpret_cast<const std::byte *>(val.data()),
+        reinterpret_cast<const std::byte *>(val.data()) + val.size());
+    scan_cache_[kv.first] = idx;
+  }
+
+  if (tx->is_aborted()) {
+    thd_mark_transaction_to_rollback(ha_thd(), 1);
+    DBUG_RETURN(false);
+  }
+
+  if (scanned_keys_.empty()) {
+    DBUG_RETURN(false);
+  }
+
+  // The proxy fetches all rows in one RPC.
+  scan_exhausted_ = true;
+
+  DBUG_RETURN(true);
+}
+
+int ha_lineairdb::rnd_next(uchar *buf) {
+  DBUG_ENTER("ha_lineairdb::rnd_next");
+  ha_statistic_increment(&System_status_var::ha_read_rnd_next_count);
+
+  if (buffer_position_ >= scanned_keys_.size()) {
+    if (scan_exhausted_) {
+      DBUG_RETURN(HA_ERR_END_OF_FILE);
+    }
+
+    if (!fetch_next_batch()) {
+      auto tx = get_transaction(ha_thd());
+      if (tx->is_aborted()) {
+        DBUG_RETURN(abort_errno(tx));
+      }
+      scan_exhausted_ = true;
+      DBUG_RETURN(HA_ERR_END_OF_FILE);
+    }
+  }
+
+  auto &key = scanned_keys_[buffer_position_];
+  auto &value = scanned_values_[buffer_position_];
+  buffer_position_++;
+
+  int error = set_fields_from_lineairdb(buf, value.data(), value.size());
+  if (error == 0) {
+    last_fetched_primary_key_ = key;
+  }
+  current_position_++;
+  DBUG_RETURN(error);
+}
+
+void ha_lineairdb::position(const uchar *) {
+  DBUG_TRACE;
+
+  if (last_fetched_primary_key_.empty()) {
+    return;
+  }
+
+  store_primary_key_in_ref(last_fetched_primary_key_);
+}
+
+int ha_lineairdb::rnd_pos(uchar *buf, uchar *pos) {
+  DBUG_TRACE;
+
+  std::string primary_key = extract_primary_key_from_ref(pos);
+
+  if (primary_key.empty()) {
+    return HA_ERR_KEY_NOT_FOUND;
+  }
+
+  // Return from scan_cache_ if available. Without this, each sorted re-read
+  // would require a separate RPC to LineairDB.
+  auto cache_it = scan_cache_.find(primary_key);
+  if (cache_it != scan_cache_.end()) {
+    auto &value = scanned_values_[cache_it->second];
+    if (set_fields_from_lineairdb(buf, value.data(), value.size())) {
+      return HA_ERR_OUT_OF_MEM;
+    }
+    last_fetched_primary_key_ = primary_key;
+    return 0;
+  }
+
+  auto tx = get_transaction(ha_thd());
+
+  if (tx->is_aborted()) {
+    return abort_errno(tx);
+  }
+
+  tx->choose_table(db_table_name);
+  auto result = tx->read(primary_key);
+
+  // read() aborts the tx on a prefetch cache miss; catch it before the empty
+  // result below reads as a missing key.
+  if (tx->is_aborted()) {
+    return abort_errno(tx);
+  }
+  if (result.first == nullptr || result.second == 0) {
+    return HA_ERR_KEY_NOT_FOUND;
+  }
+
+  if (set_fields_from_lineairdb(buf, result.first, result.second)) {
+    tx->set_status_to_abort();
+    return HA_ERR_OUT_OF_MEM;
+  }
+
+  last_fetched_primary_key_ = primary_key;
+
+  return 0;
+}
